feature: Github Actions 워크플로우 추가 [#20]

+++ .github/scripts/BeforeInstall.bash
@@ -6,6 +6,8 @@ mkdir -vp /home/ec2-user/randhand/build
 
 docker-compose down
 
-docker stop $(docker ps -a -q)
-docker rm $(docker ps -a -q)
-docker rmi $(docker images -q)
+if [[ "$(docker ps -a -q 2> /dev/null)" != ""]];
+    docker stop $(docker ps -a -q)
+    docker rm $(docker ps -a -q)
+    docker rmi $(docker images -q)
+fi
